refactor: use filter fn from codeJS
