+++ src/js/index.js
@@ -4,8 +4,7 @@ export { getAppData } from "./data";
 export { allTemplates } from "./templates";
 export allowedResults from "./allowed_results";
 export allowedMonitors from "./allowed_monitors";
-export { getModelTreeByApplication } from "./models";
-export { filterEntityPaths } from "./filters";
+export { filterModelsByApplicationParameters } from "./models";
 export const allApplications = [
     "espresso",
     "jupyterLab",
